newlib: Add nano variant configurations
This commit adds the new configuration option 'LIBC_NANO_NEWLIB' for
specifying the additional 'nano' variant of newlib to be built.

In addition, the existing newlib 'LIBC_NEWLIB_...' configs are mirrored
as 'LIBC_NANO_NEWLIB_...' in order to accommodate specifying different
build configurations for the nano variant.

For more details, refer to the issue zephyrproject-rtos/sdk-ng#152.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>