ci: Compress build artifact file
According to the `upload-artifact` documentation, while the GitHub web
UI provides `zip` format archive for downloads, the uploaded artifacts
are actually stored in the original form and only zipped upon request.

This commit updates the CI workflow to create and upload a `.tar.gz`
format artifact to keep things sane.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>