samples: zephyr: Enable GDB source highlight
This commit enables the GDB source code syntax highlight feature
through the GNU Source Highlight library.

Note that the GDB source highlight feature is not enabled for the
Xtensa targets that use the GDB 8.3.1, mainly because the source
highlighting support in this release seems to be preliminary in this
release.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>