scripts: gdb: Fix incorrect option specification
The `--with-sysroot` configuration option was incorrectly specified
through the `do_gdb_backend` function when it should be added to the
`configure` arguments.

This commit completely removes this option since it should not be
necessary and the problem it was trying to address should have been
fixed (see 911a3d473e2b27e00d89e50d0514f9ebc3baffbf).

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>