scripts: gdb: Support building with source highlight
This commit adds support for building cross GDB with the GNU Source
Highlight library.

This library is used by the GDB to provide source code syntax
highlighting feature.

Note that this commit also removes the static libstdc++ linking
restriction when the source-highlight is enabled, since there is no
technically sound reason for that to be the case and the GDB build
scripts have been patched to allow it.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>