scripts: gdb: Ignore static libstdc++ on macOS
GDB is built using the Clang compiler on the macOS, and `libstdc++`
and `libgcc` are not applicable for obvious reasons.

Since macOS provides a fairly well known and controlled execution
environment, there is little point in static linking the C++ runtime
library.

This commit updates the GDB build script to ignore the
`CT_GDB_*_STATIC_LIBSTDCXX` configuration on the macOS.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>