ci: Check out cached source archives
This commit updates the CI workflow to check out the cached source
archives from a GitHub-hosted Git repository, in order to remove any
dependencies on the third party mirror servers.

The purpose of this is to remedy the intermittent build failures seen
in the CI due to unavailability of the third party source mirrors.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>